+++ doc/source/conf.py
@@ -55,7 +55,12 @@ release = get_versions()["version"]
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ["sphinx.ext.napoleon", "sphinxcontrib.mermaid", "sphinx.ext.autosummary"]
+extensions = [
+    "sphinx.ext.napoleon",
+    "sphinxcontrib.mermaid",
+    "sphinx.ext.autosummary",
+    "sphinxcontrib.datatemplates",
+]
 
 extensions.append("sphinx.ext.todo")
 todo_include_todos = True